services/coach-agent/pyproject.toml
[project]
name = "coach-agent"
version = "0.1.0"
description = "Agentic coaching orchestration service"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.27.0",
    "sse-starlette>=2.1.3",
    "prometheus-client>=0.21.0",
    "aiosqlite>=0.20.0",
    "google-genai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "respx>=0.21.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
